adjust to doctrine commons split out persistence
mark conflict with too old doctrine common versions
