+++ composer.json
@@ -46,6 +46,8 @@
     },
     "conflict": {
         "doctrine/phpcr-odm": "<1.4",
+        "doctrine/common": "<2.10.0",
+        "doctrine/persistence": "<1.3.0",
         "phpunit/phpunit": "<6"
     },
     "autoload": {
