+++ src/Doctrine/Phpcr/IdPrefixListener.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Cmf\Bundle\RoutingBundle\Doctrine\Phpcr;
 
-use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
+use Doctrine\Persistence\Event\LifecycleEventArgs;
 
 /**
  * Doctrine PHPCR-ODM listener to tell routes what part of their id is the URL.
